添加redis模块
