+++ cache/src/main/java/com/summer/faststarter/CacheApplication.java
@@ -1,0 +1,13 @@
+package com.summer.faststarter;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CacheApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(CacheApplication.class, args);
+  }
+
+}
