+++ cache/src/main/java/com/summer/faststarter/config/RedisCacheConfig.java
@@ -1,0 +1,42 @@
+package com.summer.faststarter.config;
+
+import org.springframework.cache.annotation.EnableCaching;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.serializer.RedisSerializer;
+
+@Configuration
+@EnableCaching
+@PropertySource("classpath:redis-local.properties")
+public class RedisCacheConfig {
+
+	@Bean
+	public RedisTemplate<Object, Object> redisTemplate(RedisConnectionFactory factory) {
+		RedisTemplate<Object, Object> redisTemplate = new RedisTemplate<>();
+		RedisSerializer<String> stringSerializer = redisTemplate.getStringSerializer();
+		redisTemplate.setKeySerializer(stringSerializer);
+		redisTemplate.setHashKeySerializer(stringSerializer);
+		redisTemplate.setHashValueSerializer(stringSerializer);
+		redisTemplate.setValueSerializer(stringSerializer);
+		redisTemplate.setConnectionFactory(factory);
+		return redisTemplate;
+	}
+
+	@Bean("valueStringSerialTemplate")
+	public RedisTemplate<Object, Object> valueStringRedisTemplate(RedisConnectionFactory factory) {
+		RedisTemplate<Object, Object> redisTemplate = new RedisTemplate<>();
+		RedisSerializer<String> redisSerializer = redisTemplate.getStringSerializer();
+
+		redisTemplate.setKeySerializer(redisSerializer);
+		redisTemplate.setValueSerializer(redisSerializer);
+
+		redisTemplate.setHashKeySerializer(redisSerializer);
+		redisTemplate.setHashValueSerializer(redisSerializer);
+
+		redisTemplate.setConnectionFactory(factory);
+		return redisTemplate;
+	}
+}
