+++ cache/src/main/resources/redis-local.properties
@@ -1,0 +1,9 @@
+#本地redis spring redis config 配置redis 服务器属性
+spring.redis.hostName=127.0.0.1
+spring.redis.port=6379
+spring.redis.timeout=5000
+#spring jedis config 配置连接池属性
+spring.redis.pool.max-active=500
+spring.redis.pool.max-idle=30
+spring.redis.pool.min-idle=10
+spring.redis.pool.max-wait=4000
