+++ cache/src/test/java/com/summer/faststarter/CacheApplicationTests.java
@@ -1,0 +1,13 @@
+package com.summer.faststarter;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CacheApplicationTests {
+
+  @Test
+  void contextLoads() {
+  }
+
+}
