+++ common/src/main/java/com/summer/faststarter/CommonApplication.java
@@ -1,0 +1,12 @@
+package com.summer.faststarter;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CommonApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(CommonApplication.class, args);
+  }
+}
