--- common/src/main/java/com/summer/faststarter/common/CommonApplication.java
@@ -1,12 +1,0 @@
-package com.summer.faststarter.common;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CommonApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(CommonApplication.class, args);
-  }
-}
