+++ common/src/test/java/com/summer/faststarter/CommonApplicationTests.java
@@ -1,0 +1,11 @@
+package com.summer.faststarter;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CommonApplicationTests {
+
+  @Test
+  void contextLoads() {}
+}
