--- common/src/test/java/com/summer/faststarter/common/CommonApplicationTests.java
@@ -1,11 +1,0 @@
-package com.summer.faststarter.common;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CommonApplicationTests {
-
-  @Test
-  void contextLoads() {}
-}
