+++ datasource/src/main/java/com/summer/faststarter/DatasourceApplication.java
@@ -1,0 +1,12 @@
+package com.summer.faststarter;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DatasourceApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(DatasourceApplication.class, args);
+  }
+}
