--- datasource/src/main/java/com/summer/faststarter/datasource/DatasourceApplication.java
@@ -1,12 +1,0 @@
-package com.summer.faststarter.datasource;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DatasourceApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(DatasourceApplication.class, args);
-  }
-}
