--- datasource/src/main/java/com/summer/faststarter/datasource/mybatis/generate/dao/UserDao.java
@@ -1,96 +1,0 @@
-package com.summer.faststarter.datasource.mybatis.generate.dao;
-
-import com.summer.faststarter.datasource.mybatis.generate.model.User;
-import com.summer.faststarter.datasource.mybatis.generate.model.UserCriteria;
-import java.util.List;
-import org.apache.ibatis.annotations.Param;
-
-public interface UserDao {
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table TB_USER
-     *
-     * @mbg.generated
-     */
-    long countByCriteria(UserCriteria criteria);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table TB_USER
-     *
-     * @mbg.generated
-     */
-    int deleteByCriteria(UserCriteria criteria);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table TB_USER
-     *
-     * @mbg.generated
-     */
-    int deleteByPrimaryKey(Integer id);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table TB_USER
-     *
-     * @mbg.generated
-     */
-    int insert(User record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table TB_USER
-     *
-     * @mbg.generated
-     */
-    int insertSelective(User record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table TB_USER
-     *
-     * @mbg.generated
-     */
-    List<User> selectByCriteria(UserCriteria criteria);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table TB_USER
-     *
-     * @mbg.generated
-     */
-    User selectByPrimaryKey(Integer id);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table TB_USER
-     *
-     * @mbg.generated
-     */
-    int updateByCriteriaSelective(@Param("record") User record, @Param("example") UserCriteria criteria);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table TB_USER
-     *
-     * @mbg.generated
-     */
-    int updateByCriteria(@Param("record") User record, @Param("example") UserCriteria criteria);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table TB_USER
-     *
-     * @mbg.generated
-     */
-    int updateByPrimaryKeySelective(User record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table TB_USER
-     *
-     * @mbg.generated
-     */
-    int updateByPrimaryKey(User record);
-}
