--- datasource/src/main/java/com/summer/faststarter/datasource/mybatis/generate/model/Test.java
@@ -1,141 +1,0 @@
-package com.summer.faststarter.datasource.mybatis.generate.model;
-
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-import java.io.Serializable;
-
-@ApiModel(value="com.summer.faststarter.datasource.mybatis.generate.model.Test")
-public class Test implements Serializable {
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column TB_TEST.id
-     *
-     * @mbg.generated
-     */
-    @ApiModelProperty(value="id")
-    private Integer id;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column TB_TEST.name
-     *
-     * @mbg.generated
-     */
-    @ApiModelProperty(value="name姓名")
-    private String name;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column TB_TEST.gender
-     *
-     * @mbg.generated
-     */
-    @ApiModelProperty(value="gender性别")
-    private Byte gender;
-
-    /**
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table TB_TEST
-     *
-     * @mbg.generated
-     */
-    private static final long serialVersionUID = 1L;
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column TB_TEST.id
-     *
-     * @return the value of TB_TEST.id
-     *
-     * @mbg.generated
-     */
-    public Integer getId() {
-        return id;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column TB_TEST.id
-     *
-     * @param id the value for TB_TEST.id
-     *
-     * @mbg.generated
-     */
-    public Test setId(Integer id) {
-        this.id = id;
-        return this;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column TB_TEST.name
-     *
-     * @return the value of TB_TEST.name
-     *
-     * @mbg.generated
-     */
-    public String getName() {
-        return name;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column TB_TEST.name
-     *
-     * @param name the value for TB_TEST.name
-     *
-     * @mbg.generated
-     */
-    public Test setName(String name) {
-        this.name = name;
-        return this;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column TB_TEST.gender
-     *
-     * @return the value of TB_TEST.gender
-     *
-     * @mbg.generated
-     */
-    public Byte getGender() {
-        return gender;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column TB_TEST.gender
-     *
-     * @param gender the value for TB_TEST.gender
-     *
-     * @mbg.generated
-     */
-    public Test setGender(Byte gender) {
-        this.gender = gender;
-        return this;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table TB_TEST
-     *
-     * @mbg.generated
-     */
-    @Override
-    public String toString() {
-        StringBuilder sb = new StringBuilder();
-        sb.append(getClass().getSimpleName());
-        sb.append(" [");
-        sb.append("Hash = ").append(hashCode());
-        sb.append(", id=").append(id);
-        sb.append(", name=").append(name);
-        sb.append(", gender=").append(gender);
-        sb.append(", serialVersionUID=").append(serialVersionUID);
-        sb.append("]");
-        return sb.toString();
-    }
-}
