+++ datasource/src/main/java/com/summer/faststarter/mybatis/generate/dao/TestDao.java
@@ -1,0 +1,96 @@
+package com.summer.faststarter.mybatis.generate.dao;
+
+import com.summer.faststarter.mybatis.generate.model.Test;
+import com.summer.faststarter.mybatis.generate.model.TestCriteria;
+import java.util.List;
+import org.apache.ibatis.annotations.Param;
+
+public interface TestDao {
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table TB_TEST
+     *
+     * @mbg.generated
+     */
+    long countByCriteria(TestCriteria criteria);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table TB_TEST
+     *
+     * @mbg.generated
+     */
+    int deleteByCriteria(TestCriteria criteria);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table TB_TEST
+     *
+     * @mbg.generated
+     */
+    int deleteByPrimaryKey(Integer id);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table TB_TEST
+     *
+     * @mbg.generated
+     */
+    int insert(Test record);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table TB_TEST
+     *
+     * @mbg.generated
+     */
+    int insertSelective(Test record);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table TB_TEST
+     *
+     * @mbg.generated
+     */
+    List<Test> selectByCriteria(TestCriteria criteria);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table TB_TEST
+     *
+     * @mbg.generated
+     */
+    Test selectByPrimaryKey(Integer id);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table TB_TEST
+     *
+     * @mbg.generated
+     */
+    int updateByCriteriaSelective(@Param("record") Test record, @Param("example") TestCriteria criteria);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table TB_TEST
+     *
+     * @mbg.generated
+     */
+    int updateByCriteria(@Param("record") Test record, @Param("example") TestCriteria criteria);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table TB_TEST
+     *
+     * @mbg.generated
+     */
+    int updateByPrimaryKeySelective(Test record);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table TB_TEST
+     *
+     * @mbg.generated
+     */
+    int updateByPrimaryKey(Test record);
+}
