+++ datasource/src/main/java/com/summer/faststarter/mybatis/generate/dao/UserDao.java
@@ -1,0 +1,96 @@
+package com.summer.faststarter.mybatis.generate.dao;
+
+import com.summer.faststarter.mybatis.generate.model.User;
+import com.summer.faststarter.mybatis.generate.model.UserCriteria;
+import java.util.List;
+import org.apache.ibatis.annotations.Param;
+
+public interface UserDao {
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table TB_USER
+     *
+     * @mbg.generated
+     */
+    long countByCriteria(UserCriteria criteria);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table TB_USER
+     *
+     * @mbg.generated
+     */
+    int deleteByCriteria(UserCriteria criteria);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table TB_USER
+     *
+     * @mbg.generated
+     */
+    int deleteByPrimaryKey(Integer id);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table TB_USER
+     *
+     * @mbg.generated
+     */
+    int insert(User record);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table TB_USER
+     *
+     * @mbg.generated
+     */
+    int insertSelective(User record);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table TB_USER
+     *
+     * @mbg.generated
+     */
+    List<User> selectByCriteria(UserCriteria criteria);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table TB_USER
+     *
+     * @mbg.generated
+     */
+    User selectByPrimaryKey(Integer id);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table TB_USER
+     *
+     * @mbg.generated
+     */
+    int updateByCriteriaSelective(@Param("record") User record, @Param("example") UserCriteria criteria);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table TB_USER
+     *
+     * @mbg.generated
+     */
+    int updateByCriteria(@Param("record") User record, @Param("example") UserCriteria criteria);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table TB_USER
+     *
+     * @mbg.generated
+     */
+    int updateByPrimaryKeySelective(User record);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table TB_USER
+     *
+     * @mbg.generated
+     */
+    int updateByPrimaryKey(User record);
+}
