+++ datasource/src/main/resources/generatorConfig.xml
@@ -47,11 +47,11 @@
       <property name="nullCatalogMeansCurrent" value="true" />
     </jdbcConnection>
     <!--指定生成model的路径-->
-    <javaModelGenerator targetPackage="com.summer.faststarter.datasource.mybatis.generate.model" targetProject="src/main/java"/>
+    <javaModelGenerator targetPackage="com.summer.faststarter.mybatis.generate.model" targetProject="src/main/java"/>
     <!--指定生成mapper.xml的路径-->
     <sqlMapGenerator targetPackage="generate" targetProject="src/main/resources/mybatis"/>
     <!--指定生成mapper接口的的路径-->
-    <javaClientGenerator type="XMLMAPPER" targetPackage="com.summer.faststarter.datasource.mybatis.generate.dao"
+    <javaClientGenerator type="XMLMAPPER" targetPackage="com.summer.faststarter.mybatis.generate.dao"
       targetProject="src/main/java"/>
 
     <!--生成全部表tableName设为%-->
