+++ datasource/src/main/resources/mybatis/extend/UserExtDao.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.summer.faststarter.datasource.mybatis.extend.UserExtDao">
+<mapper namespace="com.summer.faststarter.mybatis.extend.UserExtDao">
 
   <select id="countUserNumGrater" parameterType="java.lang.Integer" resultType="java.lang.Integer">
         SELECT count(*) FROM TB_USER
