+++ datasource/src/main/resources/mybatis/generate/TestDao.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.summer.faststarter.datasource.mybatis.generate.dao.TestDao">
-  <resultMap id="BaseResultMap" type="com.summer.faststarter.datasource.mybatis.generate.model.Test">
+<mapper namespace="com.summer.faststarter.mybatis.generate.dao.TestDao">
+  <resultMap id="BaseResultMap" type="com.summer.faststarter.mybatis.generate.model.Test">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -83,7 +83,7 @@
     -->
     id, name, gender
   </sql>
-  <select id="selectByCriteria" parameterType="com.summer.faststarter.datasource.mybatis.generate.model.TestCriteria" resultMap="BaseResultMap">
+  <select id="selectByCriteria" parameterType="com.summer.faststarter.mybatis.generate.model.TestCriteria" resultMap="BaseResultMap">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -127,7 +127,7 @@
     delete from TB_TEST
     where id = #{id,jdbcType=INTEGER}
   </delete>
-  <delete id="deleteByCriteria" parameterType="com.summer.faststarter.datasource.mybatis.generate.model.TestCriteria">
+  <delete id="deleteByCriteria" parameterType="com.summer.faststarter.mybatis.generate.model.TestCriteria">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -137,7 +137,7 @@
       <include refid="Example_Where_Clause" />
     </if>
   </delete>
-  <insert id="insert" parameterType="com.summer.faststarter.datasource.mybatis.generate.model.Test">
+  <insert id="insert" parameterType="com.summer.faststarter.mybatis.generate.model.Test">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -148,7 +148,7 @@
     insert into TB_TEST (name, gender)
     values (#{name,jdbcType=VARCHAR}, #{gender,jdbcType=TINYINT})
   </insert>
-  <insert id="insertSelective" parameterType="com.summer.faststarter.datasource.mybatis.generate.model.Test">
+  <insert id="insertSelective" parameterType="com.summer.faststarter.mybatis.generate.model.Test">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -174,7 +174,7 @@
       </if>
     </trim>
   </insert>
-  <select id="countByCriteria" parameterType="com.summer.faststarter.datasource.mybatis.generate.model.TestCriteria" resultType="java.lang.Long">
+  <select id="countByCriteria" parameterType="com.summer.faststarter.mybatis.generate.model.TestCriteria" resultType="java.lang.Long">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -218,7 +218,7 @@
       <include refid="Update_By_Example_Where_Clause" />
     </if>
   </update>
-  <update id="updateByPrimaryKeySelective" parameterType="com.summer.faststarter.datasource.mybatis.generate.model.Test">
+  <update id="updateByPrimaryKeySelective" parameterType="com.summer.faststarter.mybatis.generate.model.Test">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -234,7 +234,7 @@
     </set>
     where id = #{id,jdbcType=INTEGER}
   </update>
-  <update id="updateByPrimaryKey" parameterType="com.summer.faststarter.datasource.mybatis.generate.model.Test">
+  <update id="updateByPrimaryKey" parameterType="com.summer.faststarter.mybatis.generate.model.Test">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
