+++ datasource/src/main/resources/mybatis/generate/UserDao.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.summer.faststarter.datasource.mybatis.generate.dao.UserDao">
-  <resultMap id="BaseResultMap" type="com.summer.faststarter.datasource.mybatis.generate.model.User">
+<mapper namespace="com.summer.faststarter.mybatis.generate.dao.UserDao">
+  <resultMap id="BaseResultMap" type="com.summer.faststarter.mybatis.generate.model.User">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -84,7 +84,7 @@
     -->
     id, name, gender, age
   </sql>
-  <select id="selectByCriteria" parameterType="com.summer.faststarter.datasource.mybatis.generate.model.UserCriteria" resultMap="BaseResultMap">
+  <select id="selectByCriteria" parameterType="com.summer.faststarter.mybatis.generate.model.UserCriteria" resultMap="BaseResultMap">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -128,7 +128,7 @@
     delete from TB_USER
     where id = #{id,jdbcType=INTEGER}
   </delete>
-  <delete id="deleteByCriteria" parameterType="com.summer.faststarter.datasource.mybatis.generate.model.UserCriteria">
+  <delete id="deleteByCriteria" parameterType="com.summer.faststarter.mybatis.generate.model.UserCriteria">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -138,7 +138,7 @@
       <include refid="Example_Where_Clause" />
     </if>
   </delete>
-  <insert id="insert" parameterType="com.summer.faststarter.datasource.mybatis.generate.model.User">
+  <insert id="insert" parameterType="com.summer.faststarter.mybatis.generate.model.User">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -151,7 +151,7 @@
     values (#{name,jdbcType=VARCHAR}, #{gender,jdbcType=TINYINT}, #{age,jdbcType=INTEGER}
       )
   </insert>
-  <insert id="insertSelective" parameterType="com.summer.faststarter.datasource.mybatis.generate.model.User">
+  <insert id="insertSelective" parameterType="com.summer.faststarter.mybatis.generate.model.User">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -183,7 +183,7 @@
       </if>
     </trim>
   </insert>
-  <select id="countByCriteria" parameterType="com.summer.faststarter.datasource.mybatis.generate.model.UserCriteria" resultType="java.lang.Long">
+  <select id="countByCriteria" parameterType="com.summer.faststarter.mybatis.generate.model.UserCriteria" resultType="java.lang.Long">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -231,7 +231,7 @@
       <include refid="Update_By_Example_Where_Clause" />
     </if>
   </update>
-  <update id="updateByPrimaryKeySelective" parameterType="com.summer.faststarter.datasource.mybatis.generate.model.User">
+  <update id="updateByPrimaryKeySelective" parameterType="com.summer.faststarter.mybatis.generate.model.User">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -250,7 +250,7 @@
     </set>
     where id = #{id,jdbcType=INTEGER}
   </update>
-  <update id="updateByPrimaryKey" parameterType="com.summer.faststarter.datasource.mybatis.generate.model.User">
+  <update id="updateByPrimaryKey" parameterType="com.summer.faststarter.mybatis.generate.model.User">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
