--- datasource/src/test/java/com/summer/faststarter/datasource/DatasourceApplicationTests.java
@@ -1,11 +1,0 @@
-package com.summer.faststarter.datasource;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DatasourceApplicationTests {
-
-  @Test
-  void contextLoads() {}
-}
